api: add metadata for ttrpc context
For tracing purpose, it needs to pass metadata like span id to the server.
`ttrpc-rust` should then parse the metadata to service handler by `TtrpcContext`

This commit is mainly a porting of parts of `metadata_test.go`
and `types.go` if go version, and now only working on server side.

Signed-off-by: bin <bin@hyper.sh>