Support MacOS
This PR adds MacOS support to TTRPC rust crate.

It uses conditional compilation to exclude vsock on
non Linux platforms.

Another difference is that MacOS doesn't support creation
of a descriptor with CLOEXEC atomically, so this PR adds
fcntl calls on Linux.

Note: because of this there is small chance of descriptor leak
when fork + exec happens in between.

Signed-off-by: Maksym Pavlenko <pavlenko.maksym@gmail.com>