refactor: remove Domain::AbstractUnix
Domain::AbstractUnix and Domain::Unix are almost same
execpt that they are slightly different when making address.

Both of them can be represented by Domain::Unix and we left
'@' prefix in the address after parse_addr, use this prefix
to determine whether it is abstract.

So that the code is simplified.

Signed-off-by: Tim Zhang <tim@hyper.sh>