Check for errors when connecting to client
The unwrap() causes a panic.  This can happen if the file doesn't exist or there is a small chance when mutliple clients connect quickly that the pipe returns 'All pipe instances are busy.' also causing a panic.  This allows the caller to handle these errors and retry if necessary (as is the case with the pipe instances being busy).

Signed-off-by: James Sturtevant <jstur@microsoft.com>