Closing connections properly on Windows
The current closing logic was causing a deadlock becuase an unused pipe was being returned when the close was called on the listenering.  This sets a flag so we can detect that we should not use that pipe connection which wasn't actually connected to a client but returned from the async call becuase we told it too.

Signed-off-by: James Sturtevant <jstur@microsoft.com>