MAINTAINERS: add Tim Zhang and Lei Wang as member
Tim Zhang(@Tim-Zhang) has meaningful contribution to the project and is
the maintainer of project as collaborator. He should be added as member
officially. I'd like to add him as COMMITTER.

And Lei Wang(@wllenyj) has been a active reviewer, but he is still
collaborator. He should be added as member as well. So, I'd like to add
him as REVIEWER.

Signed-off-by: Wei Fu <fuweid89@gmail.com>