add channel to get send result
Currently the send() method of stream implemented by send the value to
an unbounded channel, so even the connection is closed for a long time,
the send function still return succeed. This commit adds a channel to the
message so that we can wait until the message is truely written to the
connection.

Signed-off-by: Abel Feng <fshb1988@gmail.com>