send all pending responses before shutting down
The async implementation doesn't wait for responses to be sent to clients
when shutting down. These responses are lost if the application exists fast
enough before the write task gets to send them. That means that the
response for a request that triggers shutdown of the server might be lost
without the client knowing that the server was actually stopped.

Signed-off-by: Alexandru Matei <alexandru.matei@uipath.com>